backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic[email]==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
